Update FacebookLoginButtonRenderer.cs
When the view that have the Facebook login button displayed are poped and transitioned into another view before the button have been disposed it's going to call the OnElementChanged after the UI have left the previous view which will make the "NewElement" property to become null. Adding a null condition to it resolves the issue and does not have any other impact on the overall functionality of the renderer.